Insert hangman.py
My project is building off the concept of the party game of hangman. This project uses many concepts learned in code in place such as sequencing, iteration, loops, and using comments to explain work. This project combines those concepts with the idea of the party game of Hangman, in order to expand it and make it accessible and easy to play on computers. It uses graphics to display a real time hangman that changes as the game goes on. There are over 100 different words and 3 difficulty categories to choose from. This game suits almost anyone!